Refactor drag and drop feature
In this commit the drag and drop featured solves issues related to the
data transmission. To solve these issues the following modifications
have been introduced:

- The dragover event listener is also applied to the container, the
task-list element. This allows more movement freedom and better overall
behavior of the feature.
- It has been added a moving class to the element while it is being
dragged. The utility is twofold, since it allows to format the moving
element while identifying it for further use.
- The drop event has been replaced with the the dragend event, because
the drop event affects the target and the dragend affects the moving
element. This enables the element to move without losing the event
listeners.
- The data is no longer being transfered with the dataTransfer
functions. The data is being moved while dragging.
- A new function has been created to retrieve the element where the
moving element should be placed, using its Y positiong and half its
height.
- The dragover event for each li now places the moving element
before the target element. If the target element is null, the moving
element will be appended at the end.